% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{cutoff}
\alias{cutoff}
\title{A Cut-Off of Unified Selection Score}
\usage{
cutoff(params, nperm = 100, seed = 1)
}
\arguments{
\item{params}{The parameters used in \code{main} function. If you already run the \code{main} function, it then provides the list of params.}

\item{nperm}{The number of permutation replicates. Default is 100.}

\item{seed}{An integer value for reproducible outcome. Default is 1 that does not have special meaning.}
}
\value{
A data frame containing cut-off values of unified selection score. Each row represents the expected number (\eqn{\theta}) of falsely selected variants and each column represents a phenotype.
}
\description{
When the expected number of falsely selected variants is given by \eqn{\theta}, the corresponding cut-off of unified selection score is calculated. Then, we conclude that the variants with unified selection score greater than the obtained cut-off value are potentially associated with multiple phenotypes, and the number of falsely selected variants among them is expected to be less than \eqn{\theta}. In the data frame of the output, each row index represent the value of \eqn{\theta}, that is the first row is \eqn{\theta=1}, the second row is \eqn{\theta=2}, and so on.
}
\references{
Kim, K., Koo, J., & Sun, H. (2020). An empirical threshold of selection probability for analysis of high-dimensional correlated data. \emph{Journal of Statistical Computation and Simulation}, 90(9), 1606-1617.
}
