% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniss.R
\name{uniss}
\alias{uniss}
\title{Unified Selection Score for Multi-Trait Analysis}
\usage{
uniss(
  x,
  y,
  family,
  lambda = NULL,
  alpha = 0.1,
  B = 100,
  psub = 1,
  pf = NULL,
  std.x = TRUE,
  std.y = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{An \code{n} by \code{p} matrix of variants; Each row represents an individual, and each column a genetic variant. Can be high-dimensional data with \code{n<p}.}

\item{y}{An \code{n} by \code{q} matrix of phenotypes, each with being either continuous or categorical.}

\item{family}{A vector of distribution family, such as 'gaussian', 'binomial', 'multinomial', and etc, corresponding to phenotypes.}

\item{lambda}{A grid vector of the penalty parameter used in elastic-net penalty. It can be provided through our function 'grid.lambda'.}

\item{alpha}{Mixing proportion between ridge and lasso penalties used in elastic-net penalty. Default is 0.1.}

\item{B}{The number of resampling replicates when calculating the selection probability. Default is 100.}

\item{psub}{Subsampling proportion. When \code{psub=1.0} bootstrap resampling is applied, and otherwise subsampling without replacement is applied. Default is 1.0.}

\item{pf}{A vector of penalty to be imposed for each column of \code{x}. The 0 value indicates no penalty which leads the genetic variants always to be included in the model. On the other hand, the 1 value indicates that the penalty is imposed for each column as it was.}

\item{std.x}{Logical flag for x variable standardization. If TRUE, each variant is scaled to have zero mean and unit variance. Default is TRUE.}

\item{std.y}{Logical flag for y variable standardization. If TRUE, each phenotype is scaled to have zero mean and unit variance. Default is TRUE.}

\item{verbose}{Logical indicating whether or not to print out progress indicators.}

\item{...}{Additional arguments passing to function \code{glmnet::glmnet}.}
}
\value{
A list of fitting results containing
\item{selscore}{a matrix of which the first four columns indicate the individual selection probabilities for phenotypes and the last column the unified selection score calculated from the adjusted selection probabilities.}
\item{params}{arguments used in model fitting}
}
\description{
New statistical selection method for pleiotropic variants associated with both quantitative and qualitative traits, based on a penalized regression and a resampling approach. Details are refer to the paper in reference section.
}
\examples{

set.seed(1)
n=100; p=200; q=4

X <- replicate(p, rbinom(n,2,0.2)) #generate the SNP X=0,1,2
b <- matrix(0, p, q)
b[1:5,1:4] <- 1.0
b[6:10,1:2] <- 1.0


Z <- replicate(1, rnorm(n))
g <- matrix(0, 1, q)
g[1,1:4] <- 0.1

x <- cbind(Z, X)
beta <- rbind(g, b)

y <- x\%*\%beta + replicate(q, rnorm(n,1))
y[,2:3] <- apply(y[,2:3], 2, function(yk) ifelse(yk > median(yk), 1, 0) )
family <- c("gaussian","binomial","binomial","gaussian")
alpha <- 0.1
pf <- c(0, rep(1, p))

lambda.all <- grid.lambda(x = x, y = y,
                               family = family,
                               iter = 10,
                               seq.alpha = alpha, nlambda = 10,
                               pf=pf)
lambda <- lapply(lambda.all, function(x) seq(median(x), max(x), length.out=10) )


fit.uniss <- uniss( x = x, y = y, family = family,
                    psub = 1.0,
                    alpha = 0.1,
                    lambda = lambda,
                    B = 20, # Default is 100
                    pf = pf,
                    std.x = TRUE,
                    std.y = TRUE )

fit.cutoff <- cutoff(params = fit.uniss$params, nperm=10)

head(fit.cutoff, 10)




}
\references{
Kim, K., Jun, T., Wang, S., & Sun, H. (2022). New statistical selection method for pleiotropic variants associated with both quantitative and qualitative traits. \emph{Bioinformatics}, submitted.
}
